Intial commit: compression algorithm
This is largely a straightforward port of the algorithm from jit [1]. It
handles edge cases at the start and end of the source string, but it's
currently very copy-heavy. Most of the copying can probably be removed,
but before we optimise I'd lack to add more features and get some
benchmarks in place.

No serialisation is done at this point, we just generate copy/insert
operations so we can test the compressor.

[1]: https://github.com/jcoglan/jit/blob/master/lib/pack/xdelta.rb