Encodings for buffer sizes and copy instructions
Git deltas use two numeric encodings:

- A delta begins with the sizes of the source and target buffers,
  encoded as little-endian variable-length integers

- Copy instructions are encoded as a 32-bit offset and 24-bit size,
  combined into a single 56-bit value, written in little-endian order,
  with zero-value bytes removed, and with a leading byte whose bits
  indicate which bytes of the value have been retained

Here we add encoders to Vec<u8> for these formats. We'll define decoders
when we know more about how these values are likely to be read from data
sources.

Being able to encode these values means we'll be able to determine the
total size of a delta before serialising the whole thing to a single
Vec.